pyproject.toml
[tool.poetry]
name = "goit-web-module4"
version = "0.1.0"
description = "The first web application"
authors = ["Nataliia Tiutiunnyk <nat.tiutiunnyk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
